+++ src/main/java/de/uniks/stp24/service/game/MarketService.java
@@ -22,7 +22,7 @@ public class MarketService {
     public MarketService() {
     }
 
-    public Observable<Map<String, Integer>> getVariables(){
+    public Observable<Map<String,Double>> getVariables(){
         return this.presetsApiService.getVariables();
     }
